+++ src/sql/kql/index.ts
@@ -5,6 +5,7 @@
 export { parseKql } from './_parser';
 export { kqlToSql } from './toSql';
 export { getSuggestions, resolveValueContext } from './suggest';
+export { escapeKqlIdent } from './_escape';
 export { loadFieldValues } from './_values';
 export { KqlSyntaxError } from './_error';
 export type { KqlNode, KqlAnd, KqlOr, KqlNot, KqlIs, KqlRange } from './ast';